rework and restructure startup logic with modularity and flexibility, rework logger configuration
