+++ db/postgres.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/blutspende/bloodlab-common/config"
 	_ "github.com/jackc/pgx/v5/stdlib" /* Postgres driver with "pgx" driver name */
 	"github.com/jmoiron/sqlx"
 	"github.com/rs/zerolog/log"
@@ -28,19 +27,6 @@ type PgConfig struct {
 	ConnectionMaxLifetimeSeconds *int
 	ConnectionMaxIdleTimeSeconds *int
 	UseOpenTelemetry             bool
-}
-
-// TODO: rethink this
-func FillPgConfigBase(configuration *config.CommonConfiguration) PgConfig {
-	return PgConfig{
-		ApplicationName: configuration.ApplicationName,
-		Host:            configuration.PostgresDB.Host,
-		Port:            configuration.PostgresDB.Port,
-		User:            configuration.PostgresDB.User,
-		Pass:            configuration.PostgresDB.Pass,
-		Database:        configuration.PostgresDB.Database,
-		SSLMode:         configuration.PostgresDB.SSLMode,
-	}
 }
 
 type Postgres interface {
